pyproject.toml
[tool.poetry]
name = "tracerail-bootstrap"
version = "0.2.0"
description = "TraceRail starter stack: Temporal + Flowable DMN + LangGraph + optional Taskana"
authors = ["Esteban Sancho <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
temporalio = "^1.6.0"
langgraph = "^0.0.45"
openai = "^1.26.0"
guardrails-ai = "^0.5.0"
rich = "^13.7.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
tracerail-core = { path = "../tracerail-core", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
